refactor(module): Remodule the client
Change the client/cli to client/employee_management/cli
Adding modles in the client/employee_management
Adding utilities in the client/employee_management
