+++ main.py
@@ -2,9 +2,9 @@ import json
 
 import click
 
-from client.cli.employee_management.employee_cli import EmployeeCLI
-from client.cli.employee_management.employee_model import EmployeeModel
-from client.cli.employee_management.timezone_utils import TimezoneUtils
+from client.employee_management.cli.employee_cli import EmployeeCLI
+from client.employee_management.models.employee_model import EmployeeModel
+from client.employee_management.utilities.timezone_utils import TimezoneUtils
 from employee_management.apps.employee.repositories.employee_repository import (
     EmployeeSQLiteRepository,
 )
